restructures tests folder
